+++ ios/ReactNative/CookieManager.h
@@ -1,4 +1,4 @@
 #import <CookieManagerSpec/CookieManagerSpec.h>
-@interface CookieManager : NSObject <NativeCookieManagerSpec>
+@interface CookieManager : NativeCookieManagerSpecBase <NativeCookieManagerSpec>
 
 @end