+++ src/NativeCookieManager.ts
@@ -1,4 +1,8 @@
-import { TurboModuleRegistry, type TurboModule } from 'react-native';
+import {
+  TurboModuleRegistry,
+  type CodegenTypes,
+  type TurboModule,
+} from 'react-native';
 
 export type CookieSameSite = 'lax' | 'strict' | 'none';
 
@@ -17,7 +21,16 @@ export type Cookie = {
 
 export type Cookies = Record<string, Cookie>;
 
+export type CookieChangeStore = 'foundation' | 'webKit';
+
+export type CookieChangeEvent = {
+  store: CookieChangeStore;
+};
+
 export interface Spec extends TurboModule {
+  readonly onCookieChange: CodegenTypes.EventEmitter<CookieChangeEvent>;
+  startCookieChangeObserving(): void;
+  stopCookieChangeObserving(): void;
   setCookie(url: string, cookie: Cookie, useWebKit?: boolean): Promise<boolean>;
   setFromResponse(url: string, cookie: string): Promise<boolean>;
   getCookies(url: string, useWebKit?: boolean): Promise<Cookies>;